requirements.txt
Django==4.1
djangorestframework==3.13.1
drf-yasg==1.21.3
Faker==13.15.1
